microservices/video_onboarding_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.0
python-dotenv==1.0.1
requests==2.31.0
aiofiles==23.2.1
openai-whisper==20250625
ffmpeg-python==0.2.0
python-multipart==0.0.20
websockets==14.1
